Extract types out
